Refactored Entity to not extend collider
There were too many layers of inheritance and it did not make sense to
inherit collider. Collider is an attribute of entity but not the parent.

Right now collider x,y,width,height are used for everything. But in
future it would be possible to make it so that Entity has different x,y
and Collider would be customizable (eg. scaled or different
implementation)