Incredibly drastic changes, basically restart
The tilemap logic and client in general was just simply not designed
with multiplayer in mind, and was plagued by global variables and
designs that would be incredibly hard to untangle.

So, learning from previous things, starting to build again from a
cleaner slate and adding things from old code to get back to where I was
faster.