Entity move: improved accuracy with larger moves
If the Entity moves a distance that is longer than tileSize, check tiles
on the way incrementally.

If a tile collision is detected, handle it and stop the movement.

Basically a 90% improvement of the previous version but has some
drawbacks:
1. Still does not work flawlessly, Entity may fall through one tile
server side.
2. Moving requires a little bit more computations... But this is minimal
and ~should not~ affect performance much at all.