Duplex protocol, decoupled updates from Connection
Uses:
1. Message queue to ensure message order

2. MD5 checksum to verify data integrity

3. ACK handshakes to validate requests were received

Snapshot is now sent as a duplex message.
Example:

1. host.updates:pushDuplex(client, data)

2. After decoding the request (Connection.lua), client sends ACK request
back to host to tell that the message was received and handled.

3. After receiving the ACK request, host sends back ACK to verify the
exchange is complete. Client receives ACK, closing the exchange.

Before the exchange is closed, the requests will automatically retry if
there is no response before retry limit is reached.

Because this system is quite a lot of new code again, I decoupled all of
the update handling into new classes.

Updates class is used to handle network updates that are common for both
server and client, eg. entity position. ServerUpdates and ClientUpdates
extend this class with functionality in a way that more suits one or the
other.

Deleted some libraries, added rockspec which probably does not work yet.
But basically the project now has 2 luarocks dependencies: luasocket and
bit32. Without bit32 the performance of md5 checksum is atrocious.