Periodic snapshots, code improvements
Removed health from tiles due to design reasons

DuplexQueue now takes in client/host and sends request immediately upon
refresh / pushing new element in the head of the queue. Due to the queue based design it is very important that the
duplex communication is as fast to reply as possible.