The "chunking" update
Tilemap now embraces bump for storing the tiles. No point in having the
same data in two different collections.

This allows some cool things: easy to query items from bump for
rendering. Rendering now only gets the items that are currently visible
on the screen from bump and renders those.

Chunking

To reduce the size of duplex state data that needs to be passed from
server to client, only send the visible chunk the player is currently
located at in the level.

Basically, any resolution the game can be played at scales to this
maximum possible screen width * height. A chunk is basically (screen
resolution) ^ 2, so the player always has everything visible loaded from
the server.

When a player switches a chunk, the server automatically sends them data
for the new chunk. (Unless they were in it just a moment ago- to to
prevent too many updates in intersections, implemented ChunkHistory for
players...)

Split snapshots into multiple messages that are sent instead of one big
message. LuaSocket has a limitation of receiving only 8.1k bytes in one
message.

Added source code for "Resources" system but it's still work in progress.